Fixed some bugs and added travis-cl auto build tester Updated the README.md to reflect that

+++ src/strings.hpp
@@ -18,7 +18,7 @@ std::string randomString(const size_t &size, const std::string &dictionary, cons
     if (int(dictionary.size()) == 0)
         throw Exception("randomString expects the `dictionary` size to be bigger than 0 when asked to generate a non-null string");
 
-    if (int64_t(maxSame) * int64_t(dictionary.size()) < size)
+    if (int64_t(maxSame) * int64_t(dictionary.size()) < int64_t(size))
         throw Exception("randomString expects the `size` to be lower than the product of the `dictionary` size and the maximum number of identical characters");
 #endif
     if (maxSame >= size) {
